Geometric rules-based detection of slash piles from CHM data

This function implements a geometric, rules-based framework to identify candidate slash piles from a Canopy Height Model (CHM). The process begins by defining a vertical search space within the CHM based on user-defined height thresholds. Once candidate segments are generated via the chosen segmentation method, the function applies two distinct shape irregularity filters to refine the results. This approach ensures that the final predictions align with the expected structural characteristics of slash piles that also meet expected sizing thresholds based on minimum and maximum area expectations set by the user. Users should determine these threshold values based on pile construction prescriptions, on-site visual observations post-treatment, or prior research on typical pile dimensions for the specific treatment/forest type. This function is the primary structural pile detection process and is intended to be used prior to piles_spectral_filter() for a complete detection workflow.

Usage

piles_detect(
  chm_rast,
  seg_method = "dbscan",
  min_ht_m,
  max_ht_m,
  min_area_m2,
  max_area_m2,
  min_convexity_ratio,
  min_circularity_ratio,
  smooth_segs = T,
  outfile = NA
)

Arguments

chm_rast

A SpatRaster object representing the Canopy Height Model.

seg_method

Character string. The segmentation algorithm to use: either "watershed" or "dbscan".

min_ht_m

Numeric. The minimum expected pile height.

max_ht_m

Numeric. The upper limit of the vertical search space and expected pile height.

min_area_m2

Numeric. The minimum pile footprint area required for a segment to be retained as a candidate.

max_area_m2

Numeric. The maximum pile footprint area expected.

min_convexity_ratio

Numeric (0 to 1). Also known as the "solidity" ratio. Calculated as (Area of Polygon / Area of Convex Hull). This metric identifies segments with deep indents, holes, or branching. A perfectly convex shape like a circle or square has a ratio of 1.0. Values closer to 0 indicate high irregularity or concavity. Note that convexity is not sensitive to elongation.

min_circularity_ratio

Numeric (0 to 1). Also known as the "Reock Compactness Score." Calculated as (Area of Polygon / Area of Minimum Bounding Circle). This measures how closely a shape is spread around its central point. A perfect circle has a ratio of 1.0. For reference, a perfect square is approximately 0.637 and an equilateral triangle is approximately 0.414.

smooth_segs

Logical. Should convex hull smoothing be applied to the raster-detected segment boundaries to reduce pixelated edges? Any resulting smoothed segments that overlap are removed based since slash piles are generally distinct objects on the landscape.

outfile

Character string or NA. If NA (default), the function returns the segmented piles as an sf object. If a valid file path is provided (e.g., "./output/detected_piles.gpkg"), the result is written to a GeoPackage file and the function returns the character string of the file path.

Value

If outfile is NA, returns an sf object of candidate pile polygons. If outfile is a character string, returns the character string of the saved file path.

Examples

if (FALSE) { # \dontrun{
 ###########################################
 # piles_detect() test
 ###########################################
 # load a chm
 chm_fnm <- system.file(package = "cloud2trees", "extdata", "piles_chm.tif")
 my_chm <- terra::rast(chm_fnm)
 terra::plot(my_chm, axes = F)
 # piles_detect() that
 piles_detect_ans <- piles_detect(
   chm_rast = my_chm
   , seg_method = "dbscan"
   , min_ht_m = 1
   , max_ht_m = 4
   , min_area_m2 = pi*(1.5/2)^2
   , max_area_m2 = pi*(6/2)^2
   , min_convexity_ratio = 0.3
   , min_circularity_ratio = 0.4
   , smooth_segs = T
   , outfile = NA
 )
 # what?
 piles_detect_ans
 dplyr::glimpse(piles_detect_ans$segs_sf) # here are pile polygons
 # plot on orig chm
 terra::plot(my_chm, axes = F)
 piles_detect_ans$segs_sf %>%
   terra::vect() %>%
   terra::plot(border = "magenta", lwd = 2, col = NA, add = T)
 # plot on chm slice generated by piles_detect()
 terra::plot(piles_detect_ans$slice_chm_rast, axes = F, col = viridis::cividis(n=100))
 piles_detect_ans$segs_sf %>%
   terra::vect() %>%
   terra::plot(border = "magenta", lwd = 2, col = NA, add = T)
 # what if we use the `outfile` arg in piles_detect() ?
 my_fnm <- tempfile()
 piles_detect_ans2 <- piles_detect(
   chm_rast = my_chm
   , seg_method = "dbscan"
   , min_ht_m = 1
   , max_ht_m = 6.6
   , min_area_m2 = pi*(1.5/2)^2
   , max_area_m2 = pi*(8/2)^2
   , min_convexity_ratio = 0.11
   , min_circularity_ratio = 0.22
   , smooth_segs = T
   , outfile = my_fnm
 )
 piles_detect_ans2 # this is now a filename
 piles <- sf::st_read(piles_detect_ans2) # read it
 # plot on orig chm
 terra::plot(my_chm, axes = F)
 piles %>%
   terra::vect() %>%
   terra::plot(border = "magenta", lwd = 2, col = NA, add = T)
} # }