Spectral filtering of candidate slash piles

Evaluates candidate segments generated by piles_detect() against an RGB raster. It uses a voting system based on six spectral indices to separate woody slash material from background vegetation or ground. This function can either return a filtered dataset or the full input dataset with appended spectral metrics for diagnostic purposes.

Usage

piles_spectral_filter(
  sf_data,
  rgb_rast,
  red_band_idx,
  green_band_idx,
  blue_band_idx,
  spectral_weight = 4,
  filter_return = T
)

Arguments

sf_data

An sf object of candidate polygons to be filtered.

rgb_rast

A multi-band SpatRaster containing RGB imagery.

red_band_idx

Integer. Index of the red band.

green_band_idx

Integer. Index of the green band.

blue_band_idx

Integer. Index of the blue band.

spectral_weight

Integer (0 to 6). The consensus threshold requiring a specific number of index thresholds to be met for a candidate pile to be retained.

filter_return

Logical. If TRUE (default), the function returns only polygons where the "inrange_th_votes" is greater than or equal to the spectral_weight. If FALSE, the function returns the entire input dataset with columns for each calculated spectral index and the "inrange_th_votes" column for post-processing and analysis.

Value

An sf object. Depending on filter_return, this will either be a subset of the input data or the full input data with appended spectral voting columns.

Examples

if (FALSE) { # \dontrun{
 ###########################################
 # piles_detect() test
 ###########################################
 # load a chm
 chm_fnm <- system.file(package = "cloud2trees", "extdata", "piles_chm.tif")
 my_chm <- terra::rast(chm_fnm)
 terra::plot(my_chm, axes = F)
 # piles_detect() that
 piles_detect_ans <- piles_detect(
   chm_rast = my_chm
   , seg_method = "dbscan"
   , min_ht_m = 1
   , max_ht_m = 4
   , min_area_m2 = pi*(1.5/2)^2
   , max_area_m2 = pi*(6/2)^2
   , min_convexity_ratio = 0.3
   , min_circularity_ratio = 0.4
   , smooth_segs = T
   , outfile = NA
 )
 # what?
 piles_detect_ans
 dplyr::glimpse(piles_detect_ans$segs_sf) # here are pile polygons
 # plot on orig chm
 terra::plot(my_chm, axes = F)
 piles_detect_ans$segs_sf %>%
   terra::vect() %>%
   terra::plot(border = "magenta", lwd = 2, col = NA, add = T)
 # plot on chm slice generated by piles_detect()
 terra::plot(piles_detect_ans$slice_chm_rast, axes = F, col = viridis::cividis(n=100))
 piles_detect_ans$segs_sf %>%
   terra::vect() %>%
   terra::plot(border = "magenta", lwd = 2, col = NA, add = T)
 # what if we use the `outfile` arg in piles_detect() ?
 my_fnm <- tempfile()
 piles_detect_ans2 <- piles_detect(
   chm_rast = my_chm
   , seg_method = "dbscan"
   , min_ht_m = 1
   , max_ht_m = 6.6
   , min_area_m2 = pi*(1.5/2)^2
   , max_area_m2 = pi*(8/2)^2
   , min_convexity_ratio = 0.11
   , min_circularity_ratio = 0.22
   , smooth_segs = T
   , outfile = my_fnm
 )
 piles_detect_ans2 # this is now a filename
 piles <- sf::st_read(piles_detect_ans2) # read it
 # plot on orig chm
 terra::plot(my_chm, axes = F)
 piles %>%
   terra::vect() %>%
   terra::plot(border = "magenta", lwd = 2, col = NA, add = T)
 ###########################################
 # piles_spectral_filter() test
 ###########################################
 # load rgb data
 rgb_fnm <- system.file(package = "cloud2trees", "extdata", "piles_rgb.tif")
 my_rgb <- terra::rast(rgb_fnm)
 my_rgb
 terra::plotRGB(my_rgb)
 piles %>%
   terra::vect() %>%
   terra::plot(border = "magenta", lwd = 2, col = NA, add = T)
 # piles_detect() that
 piles_spectral_filter_ans <- piles_spectral_filter(
   sf_data = piles
   , rgb_rast = my_rgb
   , red_band_idx = 1
   , green_band_idx = 2
   , blue_band_idx = 3
   , spectral_weight = 5
   , filter_return = F # don't filter the return so we can see what did not pass
 )
 piles_spectral_filter_ans
 # look at all these indices
 piles_spectral_filter_ans$rgb_indices_rast %>%
   terra::plot(
     nc = 5
     , col = grDevices::gray.colors(111, start = 0, end = 1)
     , mar = c(0.2,0.2,1.6,0.2)
     , axes = FALSE
     , legend = F
   )
 # look at the data added to the piles
 piles_spectral_filter_ans$segs_sf %>% dplyr::glimpse()
 # how many would have been filtered if
 # ... we set `filter_return = T`
 # ... with `spectral_weight = 5` ?
 piles_spectral_filter_ans$segs_sf %>%
   sf::st_drop_geometry() %>%
   dplyr::mutate(is_spectrally_removed = inrange_th_votes<5) %>%
   dplyr::count(is_spectrally_removed)
 # plot it
 terra::plotRGB(
   my_rgb
   , main = "candidate piles kept (magenta) or removed (red)"
   , mar = c(0.2,0.2,2,0.2)
 )
 piles_spectral_filter_ans$segs_sf %>%
   dplyr::filter(inrange_th_votes>=5) %>%
   terra::vect() %>%
   terra::plot(border = "magenta", lwd = 2, col = NA, add = T)
 piles_spectral_filter_ans$segs_sf %>%
   dplyr::filter(inrange_th_votes<5) %>%
   terra::vect() %>%
   terra::plot(border = "red", lwd = 2, col = NA, add = T)
} # }